+++ src/js/calculation.js
@@ -19,10 +19,16 @@ module.exports = {
 
         updated = updated || {};
 
-        $nodes = this.form.getRelatedNodes( 'data-calculate', '', updated );
-
-        // add relevant items that have a (any) calculation
-        $nodes = $nodes.add( this.form.getRelatedNodes( 'data-relevant', '[data-calculate]', updated ) );
+        if ( updated.relevantPath ) {
+            // Questions that are descendants of a group:
+            $nodes = this.form.getRelatedNodes( 'data-calculate', '[name^="' + updated.relevantPath + '/"]' )
+                // Individual questions:
+                .add( this.form.getRelatedNodes( 'data-calculate', '[name="' + updated.relevantPath + '"]' ) )
+                // Individual radiobutton questions with a calculate....:
+                .add( this.form.getRelatedNodes( 'data-calculate', '[data-name="' + updated.relevantPath + '"]' ) );
+        } else {
+            $nodes = this.form.getRelatedNodes( 'data-calculate', '', updated );
+        }
 
         $nodes.each( function() {
             var result;
@@ -48,7 +54,6 @@ module.exports = {
             // for inputs that have a calculation and need to be validated
             constraintExpr = that.form.input.getConstraint( $this );
             relevantExpr = that.form.input.getRelevant( $this );
-
             dataNodesObj = that.form.model.node( name );
             dataNodes = dataNodesObj.get();
 
@@ -71,7 +76,44 @@ module.exports = {
             }
 
             function updateCalc( index ) {
-                relevant = ( relevantExpr ) ? that.form.model.evaluate( relevantExpr, 'boolean', name, index ) : true;
+                var pathParts = name.split( '/' );
+                /*
+                 * First determine immediate group parent of node, which will always be in correct location in DOM. This is where
+                 * we can use the index to be guaranteed to get the correct node.
+                 * (also for nodes in #calculated-items).
+                 *
+                 * Then get all the group parents of that node.
+                 * 
+                 * TODO: determine index at every level to proerly support repeats and nested repeats
+                 * 
+                 * Note: getting the parents of $this wouldn't work for nodes inside #calculated-items!
+                 */
+                var parentPath = pathParts.splice( 0, pathParts.length - 1 ).join( '/' );
+                var $parentGroups = that.form.view.$.find( '.or-group[name="' + parentPath + '"],.or-group-data[name="' + parentPath + '"]' ).eq( index )
+                    .parents( '.or-group, .or-group-data' ).addBack();
+
+                if ( $parentGroups.length ) {
+                    // Start at the highest level, and traverse down to the DOM to the immediate parent group.
+                    relevant = $parentGroups.filter( '[data-relevant]' ).reverse().get().map( function( group ) {
+                        var $group = $( group );
+                        var nm = that.form.input.getName( $group );
+
+                        return {
+                            context: nm,
+                            // thankfully relevants on repeats are not possible with XLSForm-produced forms
+                            index: that.form.view.$.find( '.or-group[name="' + nm + '"], .or-group-data[name="' + nm + '"]' ).index( $group ), // performance....
+                            expr: that.form.input.getRelevant( $group )
+                        };
+                    } ).concat( [ {
+                        context: name,
+                        index: index,
+                        expr: relevantExpr
+                    } ] ).every( function( item ) {
+                        return ( item.expr ) ? that.form.model.evaluate( item.expr, 'boolean', item.context, item.index ) : true;
+                    } );
+                } else {
+                    relevant = ( relevantExpr ) ? that.form.model.evaluate( relevantExpr, 'boolean', name, index ) : true;
+                }
 
                 // not sure if using 'string' is always correct
                 newExpr = that.form.replaceChoiceNameFn( expr, 'string', name, index );